+++ html/offscreen.html
@@ -4,9 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Offscreen FFmpeg Worker</title>
-    <script src="../lib/ffmpeg.min.js"></script>
-    <script src="../lib/ffmpeg-core.js"></script>
-    <script type="module" src="../js/offscreen.js"></script>
+    <script type="module" src="../js/runFFmpeg.js"></script>
 </head>
 <body>
 </body>
